pyproject.toml
[tool.poetry]
name = "nexus"
version = "0.1.0"
description = "A learning portal that models courses as concept maps."
authors = ["Alisha Manocha <alisha.manocha04@gmail.com>"]
readme = "README.md"
packages = [
    { include = "backend" }
]

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = "^0.117.1"
uvicorn = "^0.36.0"
pymongo = "^4.15.1"
networkx = "^3.5"
pydantic = { version = "^2.11.9", extras = ["email"] }
mypy = "^1.18.2"
ruff = "^0.13.1"
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.9.0"
isort = "^6.0.1"

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
warn_redundant_casts = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "D", "B", "I"]
ignore = ["E203", "E501", "D203", "D213", "B008"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
